CrudRepository generic part was edited.(<T,Integer> to convert <T,Long>)

+++ src/main/java/com/syrisa/springlibrarydocker/repository/AuthorRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AuthorRepository extends CrudRepository<Author, Integer> {
+public interface AuthorRepository extends CrudRepository<Author, Long> {
     Author findAuthorByAuthorName(String authorName);
 
     Page<Author> findAll(Pageable pageable);
